Fix connection rejection.
