+++ collectors/jmx/src/main/java/com/groupon/lex/metrics/jmx/MetricListener.java
@@ -1,21 +1,21 @@
 /*
  * Copyright (c) 2016, Groupon, Inc.
- * All rights reserved. 
+ * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
- * are met: 
+ * are met:
  *
  * Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer. 
+ * this list of conditions and the following disclaimer.
  *
  * Redistributions in binary form must reproduce the above copyright
  * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution. 
+ * documentation and/or other materials provided with the distribution.
  *
  * Neither the name of GROUPON nor the names of its contributors may be
  * used to endorse or promote products derived from this software without
- * specific prior written permission. 
+ * specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
  * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -31,42 +31,241 @@
  */
 package com.groupon.lex.metrics.jmx;
 
-import javax.management.MXBean;
+import com.groupon.lex.metrics.GroupGenerator;
+import com.groupon.lex.metrics.MetricGroup;
+import com.groupon.lex.metrics.Tags;
+import com.groupon.lex.metrics.jmx.JmxClient.ConnectionDecorator;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collection;
+import static java.util.Collections.singleton;
+import static java.util.Collections.unmodifiableList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.CancellationException;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.management.InstanceNotFoundException;
+import javax.management.ListenerNotFoundException;
+import javax.management.MBeanServerConnection;
+import javax.management.MBeanServerDelegate;
+import javax.management.MBeanServerNotification;
+import javax.management.Notification;
+import javax.management.NotificationListener;
 import javax.management.ObjectName;
+import lombok.Getter;
+import static java.util.Objects.requireNonNull;
 
 /**
+ * A JMX bean listener, that detects new MBeans with a given name pattern and
+ * registers them as JMX based metrics.
  *
  * @author ariane
  */
-@MXBean
-public interface MetricListener {
-    /**
-     * Check if automatic detection is enabled.
-     * @return True if the listener is enabled.
-     */
-    public boolean isEnabled();
+public class MetricListener implements GroupGenerator {
+    private final Collection<ObjectName> filter_;
+    private final Map<ObjectName, MBeanGroup> detected_groups_ = new HashMap<>();
+    private boolean is_enabled_ = false;
+    @Getter
+    private final JmxClient connection;
+    private final ConnectionDecorator decorator_;
+    private final NotificationListener listener_;
+    @Getter
+    private final List<String> subPath;
+    @Getter
+    private final Tags tags;
+    private int failCount = 0;
+
+    public MetricListener(JmxClient conn, Collection<ObjectName> filter, List<String> sub_path, Tags tags) throws IOException, InstanceNotFoundException {
+        if (filter.isEmpty())
+            throw new IllegalArgumentException("empty filter");
+        filter_ = filter;
+        connection = conn;
+        subPath = unmodifiableList(new ArrayList<>(sub_path));
+        this.tags = requireNonNull(tags);
+
+        listener_ = (Notification notification, Object handback) -> {
+            MBeanServerNotification mbs = (MBeanServerNotification) notification;
+
+            if (!filter_.stream()
+                    .anyMatch((f) -> f.apply(mbs.getMBeanName()))) {
+                Logger.getLogger(getClass().getName()).log(Level.FINER, "notification for {0} ignored: does not match filter {1}", new Object[]{mbs.getMBeanName(), filter_});
+                return;
+            }
+
+            if (null != mbs.getType()) {
+                switch (mbs.getType()) {
+                    case MBeanServerNotification.REGISTRATION_NOTIFICATION:
+                        Logger.getLogger(getClass().getName()).log(Level.INFO, "MBean Registered [{0}]", mbs.getMBeanName());
+                        onNewMbean(mbs.getMBeanName());
+                        break;
+                    case MBeanServerNotification.UNREGISTRATION_NOTIFICATION:
+                        Logger.getLogger(getClass().getName()).log(Level.INFO, "MBean Unregistered [{0}]", mbs.getMBeanName());
+                        onRemovedMbean(mbs.getMBeanName());
+                        break;
+                }
+            }
+        };
+
+        decorator_ = (MBeanServerConnection mbsc) -> {
+            /* Clear gathered metrics, since they'll belong to the old connection and won't function properly. */
+            Collection<ObjectName> allNames = new ArrayList<ObjectName>() {
+                {
+                    addAll(detected_groups_.keySet());
+                }
+            };
+            allNames.forEach(this::onRemovedMbean);
+
+            /* Initialize listener for new MBean registrations. */
+            try {
+                mbsc.addNotificationListener(MBeanServerDelegate.DELEGATE_NAME, listener_, null, null);
+            } catch (InstanceNotFoundException ex) {
+                Logger.getLogger(MetricListener.class.getName()).log(Level.SEVERE, "your MBeanServer is not compliant", ex);
+                throw new IOException("your MBeanServer is not compliant", ex);
+            }
+
+            /* Add all objects that the listener misses, due to them existing already. */
+            for (ObjectName f : filter_) {
+                mbsc.queryNames(f, null)
+                        .forEach((objname) -> onNewMbean(objname));
+            }
+        };
+    }
 
     /**
-     * Returns a list of detected names.
-     * @return The object names that were detected by this listener.
+     * Respond to MBeans being added.
+     *
+     * @param obj The name of the MBean being added.
      */
-    public ObjectName[] getDetectedNames();
+    private synchronized void onNewMbean(ObjectName obj) {
+        if (detected_groups_.keySet().contains(obj)) {
+            Logger.getLogger(MetricListener.class.getName()).log(Level.WARNING, "skipping registration of {0}: already present", obj);
+            return;
+        }
+
+        MBeanGroup instance = new MBeanGroup(connection, obj, subPath, tags);
+        detected_groups_.put(obj, instance);
+        Logger.getLogger(MetricListener.class.getName()).log(Level.INFO, "registered metrics for {0}: {1}", new Object[]{obj, instance});
+    }
 
     /**
-     * The filter for the MBeans this listener will detect.
-     * @return The object name pattern for this listener (may contain wildcards).
+     * Respond to MBeans being removed.
+     *
+     * @param obj The name of the MBean being removed.
      */
-    public ObjectName[] getFilter();
+    private synchronized void onRemovedMbean(ObjectName obj) {
+        if (!detected_groups_.keySet().contains(obj)) {
+            Logger.getLogger(MetricListener.class.getName()).log(Level.WARNING, "skipping de-registration of {0}: not present", obj);
+            return;
+        }
 
-    /**
-     * Enable the metric listener.
-     * @throws java.lang.Exception if enabling the listener failed.
-     */
-    public void enable() throws Exception;
+        MBeanGroup instance = detected_groups_.get(obj);
+        detected_groups_.remove(obj);
+        Logger.getLogger(MetricListener.class.getName()).log(Level.INFO, "de-registered metrics for {0}: {1}", new Object[]{obj, instance});
+    }
 
-    /**
-     * Disable the metric listener.
-     * @throws java.lang.Exception if disabling the listener failed.
-     */
-    public void disable() throws Exception;
+    public synchronized void enable() throws IOException {
+        if (is_enabled_)
+            return;
+
+        connection.addRecoveryCallback(decorator_);
+        is_enabled_ = true;
+        Logger.getLogger(MetricListener.class.getName()).log(Level.INFO, "enabled");
+    }
+
+    public synchronized void disable() throws IOException {
+        if (!is_enabled_)
+            return;
+
+        is_enabled_ = false;
+
+        try {
+            Optional<MBeanServerConnection> optionalConnection = connection.getOptionalConnection();
+            try {
+                if (optionalConnection.isPresent())
+                    optionalConnection.get().removeNotificationListener(MBeanServerDelegate.DELEGATE_NAME, listener_);
+            } catch (InstanceNotFoundException ex) {
+                Logger.getLogger(MetricListener.class.getName()).log(Level.SEVERE, "MBean Platform server has mysteriously disappeared...", ex);
+                throw new IllegalStateException("MBean Platform server has mysteriously disappeared", ex);
+            } catch (ListenerNotFoundException ex) {
+                Logger.getLogger(MetricListener.class.getName()).log(Level.SEVERE, "Listener was not found...", ex);
+            }
+        } finally {
+            /* Remove gathered items even if connection logic fails. */
+            Collection<ObjectName> allNames = new ArrayList<ObjectName>() {
+                {
+                    addAll(detected_groups_.keySet());
+                }
+            };
+            allNames.forEach(this::onRemovedMbean);
+
+            Logger.getLogger(MetricListener.class.getName()).log(Level.INFO, "disabled");
+        }
+    }
+
+    public ObjectName[] getFilter() {
+        return filter_.toArray(new ObjectName[0]);
+    }
+
+    public boolean isEnabled() {
+        return is_enabled_;
+    }
+
+    public ObjectName[] getDetectedNames() {
+        return detected_groups_.keySet().toArray(new ObjectName[0]);
+    }
+
+    @Override
+    public void close() throws IOException {
+        disable();
+    }
+
+    @Override
+    public Collection<CompletableFuture<Collection<MetricGroup>>> getGroups(ExecutorService executor, CompletableFuture<?> timeout) {
+        CompletableFuture<Collection<MetricGroup>> future
+                = CompletableFuture.supplyAsync(() -> readMetricGroups(timeout), executor);
+
+        timeout.whenCompleteAsync((ignoredValue, ignoredExc) -> {
+            if (!future.isDone()) {
+                future.cancel(true);
+                synchronized (this) {
+                    if (++failCount > 5) {
+                        failCount = 0;
+                        connection.rejectCurrentConnection(executor);
+                    }
+                }
+            }
+        }, executor);
+
+        return singleton(future);
+    }
+
+    private synchronized Collection<MetricGroup> readMetricGroups(CompletableFuture<?> timeout) {
+        try {
+            /*
+             * Force the connection to open, even if there are no groups to scan.
+             *
+             * If there are no groups registered, the connection opening won't be triggered otherwise.
+             * And if that isn't triggered, registering groups won't happen either.
+             */
+            connection.getConnection();
+        } catch (IOException ex) {
+            /*
+             * Connection down, can't collect any data.
+             */
+            throw new RuntimeException(ex.getMessage(), ex);
+        }
+
+        final List<MetricGroup> groups = new ArrayList<>(detected_groups_.size());
+        for (MBeanGroup group : detected_groups_.values()) {
+            if (timeout.isDone())
+                throw new CancellationException("timed out");
+            group.getMetrics().ifPresent(groups::add);
+        }
+        return groups;
+    }
 }
